Increase speed of horqrux sampling:

New:
In [3]: %timeit sample(RX(20,'theta'),backend='horqrux')
45.3 ms ± 187 μs per loop (mean ± std. dev. of 7 runs, 10 loops each)

Old:
In [2]: %timeit sample(RX(20,'theta'),backend='horqrux')
4.28 s ± 25.5 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)