@atiyo 2024-06-05
> @atiyo do you see a nice way to do this even better?

Looks good to me.

Only thing that comes to mind is trying to optimise that dictionary comprehension into something Jax native. e.g. vmapping that as well

 But maybe that's a bottleneck at all, in which case it looks good as is to me.